Split actions.js into separate action files for each page
